Fixes for re-enabling more MSVC level 4 warnings: ppapi/ edition
This contains fixes for the following sorts of issues:
* Assignment inside conditional
* Possibly-uninitialized local variable
* Signedness mismatch

This also contains a small number of other cleanups/simplifications to nearby
code.

BUG=81439
TEST=none
R=teravest@chromium.org

Review URL: https://codereview.chromium.org/375133003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@282133 0039d316-1c4b-4281-b951-d872f2087c98

+++ ppapi/tests/test_file_mapping.cc
@@ -304,7 +304,7 @@ std::string TestFileMapping::TestBadParameters() {
           instance_->pp_instance(),
           file_io.pp_resource(),
           page_size,
-          ~PP_FILEMAPPROTECTION_READ,
+          ~static_cast<uint32_t>(PP_FILEMAPPROTECTION_READ),
           PP_FILEMAPFLAG_PRIVATE,
           0,
           &address,
@@ -350,7 +350,7 @@ std::string TestFileMapping::TestBadParameters() {
           file_io.pp_resource(),
           page_size,
           PP_FILEMAPPROTECTION_READ,
-          ~PP_FILEMAPFLAG_SHARED,
+          ~static_cast<uint32_t>(PP_FILEMAPFLAG_SHARED),
           0,
           &address,
           callback.GetCallback().pp_completion_callback()));
@@ -365,7 +365,7 @@ std::string TestFileMapping::TestBadParameters() {
           file_io.pp_resource(),
           page_size,
           PP_FILEMAPPROTECTION_READ,
-          ~PP_FILEMAPFLAG_SHARED,
+          ~static_cast<uint32_t>(PP_FILEMAPFLAG_SHARED),
           1,
           &address,
           callback.GetCallback().pp_completion_callback()));
